25/06/21

+++ application/models/M_Proses_akuntansi.php
@@ -2,43 +2,28 @@
 
 	class M_Proses_akuntansi extends CI_model {
 
-		public function getPerMasaTahun($bulan, $tahun) { // not used
-			return $this->db->from('proses_akuntansi')
-							->join('tugas', 'tugas.id_tugas = proses_akuntansi.id_tugas', 'left')
-							->join('pengiriman_akuntansi', 'proses_akuntansi.id_kirim = pengiriman_akuntansi.id_pengiriman', 'left')
-							->join('permintaan_akuntansi', 'permintaan_akuntansi.id_permintaan = pengiriman_akuntansi.id_permintaan', 'left')
-							->join('jenis_data', 'permintaan_akuntansi.kode_jenis = jenis_data.kode_jenis', 'left')
-							->join('klien', 'permintaan_akuntansi.id_klien = klien.id_klien', 'left')
-							->join('user', 'proses_akuntansi.id_akuntan = user.id_user', 'left')
-							->where(['masa'=>$bulan, 'tahun'=>$tahun])
-							->order_by('id_proses', 'ASC')
-							->get()->result_array();
-		}
-
-		public function getProses($start='', $limit='', $status, $bulan, $tahun, $klien='') {
+		public function getByMasa($status, $bulan, $tahun, $klien='', $start, $limit) {
 			if($klien) {
 				$this->db->where_in('permintaan_akuntansi.id_klien', $klien);
 			}
 			if($status == 'belum') {
-				$this->db->where(['id_proses'=>null]);
-			} elseif($status == 'selesai') {
-				$this->db->where(['id_proses !='=>null, 'tanggal_selesai !='=>null]);
+				$this->db->where(['tanggal_mulai'=>null]);
 			} elseif($status == 'onproses') {
-				$this->db->where(['id_proses !='=>null, 'tanggal_selesai'=>null]);
-			}
-			if($limit) {
-				$this->db->limit($limit, $start);
+				$this->db->where(['tanggal_mulai !='=>null, 'tanggal_selesai'=>null]);
+			} elseif($status == 'selesai') {
+				$this->db->where(['tanggal_mulai !='=>null, 'tanggal_selesai !='=>null]);
 			}
 			
-			return $this->db->from('pengiriman_akuntansi')
-							->join('permintaan_akuntansi', 'id_permintaan', 'left')
-							->join('jenis_data', 'kode_jenis', 'left')
-							->join('klien', 'id_klien', 'left')
-							->join('tugas', 'jenis_data.kode_jenis = tugas.kode_jenis AND klien.status_pekerjaan = tugas.status_pekerjaan', 'left')
-							->join('proses_akuntansi', 'proses_akuntansi.id_kirim = pengiriman_akuntansi.id_pengiriman AND tugas.id_tugas = proses_akuntansi.id_tugas', 'left')
-							->join('user', 'proses_akuntansi.id_akuntan = user.id_user', 'left')
-							->where(['masa'=>$bulan, 'tahun'=>$tahun])
-							->order_by('pengiriman_akuntansi.id_pengiriman', 'ASC')
+			return $this->db->from('data_akuntansi')
+							->join('proses_akuntansi', 'proses_akuntansi.id_proses = data_akuntansi.id_kerja', 'left')
+							->join('permintaan_akuntansi', 'permintaan_akuntansi.id_permintaan = data_akuntansi.id_request', 'left')
+							->join('pengiriman_akuntansi', 'pengiriman_akuntansi.id_pengiriman = data_akuntansi.id_kirim', 'left')
+							->join('jenis_data', 'jenis_data.kode_jenis = data_akuntansi.id_jenis', 'left')
+							->join('klien', 'klien.id_klien = permintaan_akuntansi.id_klien', 'left')
+							->join('tugas', 'tugas.id_jenis = jenis_data.kode_jenis AND tugas.status_pekerjaan = klien.status_pekerjaan', 'left')
+							->join('user', 'user.id_user = proses_akuntansi.id_akuntan', 'left')
+							->where(['status'=>3, 'bulan'=>$bulan, 'tahun'=>$tahun])
+							->order_by('id_proses', 'ASC')
 							->get()->result_array();
 		}
 
@@ -47,93 +32,63 @@
 				$this->db->where_in('permintaan_akuntansi.id_klien', $klien);
 			}
 			if($status == 'belum') {
-				$this->db->where(['id_proses'=>null]);
+				$this->db->where(['tanggal_mulai'=>null]);
+			} elseif($status == 'onproses') {
+				$this->db->where(['tanggal_mulai !='=>null, 'tanggal_selesai'=>null]);
 			} elseif($status == 'selesai') {
-				$this->db->where(['id_proses !='=>null, 'tanggal_selesai !='=>null]);
-			} else {
-				$this->db->where(['id_proses !='=>null, 'tanggal_selesai'=>null]);
+				$this->db->where(['tanggal_mulai !='=>null, 'tanggal_selesai !='=>null]);
 			}
-			return $this->db->from('pengiriman_akuntansi')
-							->join('permintaan_akuntansi', 'permintaan_akuntansi.id_permintaan = pengiriman_akuntansi.id_permintaan', 'left')
-							->join('jenis_data', 'permintaan_akuntansi.kode_jenis = jenis_data.kode_jenis', 'left')
-							->join('klien', 'permintaan_akuntansi.id_klien = klien.id_klien', 'left')
-							->join('tugas', 'jenis_data.kode_jenis = tugas.kode_jenis AND klien.status_pekerjaan = tugas.status_pekerjaan', 'left')
-							->join('proses_akuntansi', 'proses_akuntansi.id_kirim = pengiriman_akuntansi.id_pengiriman AND tugas.id_tugas = proses_akuntansi.id_tugas', 'left')
-							->join('user', 'proses_akuntansi.id_akuntan = user.id_user', 'left')
-							->where(['masa'=>$bulan, 'tahun'=>$tahun])
+			
+			return $this->db->from('data_akuntansi')
+							->join('proses_akuntansi', 'proses_akuntansi.id_proses = data_akuntansi.id_kerja', 'left')
+							->join('permintaan_akuntansi', 'permintaan_akuntansi.id_permintaan = data_akuntansi.id_request', 'left')
+							->join('pengiriman_akuntansi', 'pengiriman_akuntansi.id_pengiriman = data_akuntansi.id_kirim', 'left')
+							->join('jenis_data', 'jenis_data.kode_jenis = data_akuntansi.id_jenis', 'left')
+							->join('klien', 'klien.id_klien = permintaan_akuntansi.id_klien', 'left')
+							->join('tugas', 'tugas.id_jenis = jenis_data.kode_jenis AND tugas.status_pekerjaan = klien.status_pekerjaan', 'left')
+							->join('user', 'user.id_user = proses_akuntansi.id_akuntan', 'left')
+							->where(['status'=>3, 'bulan'=>$bulan, 'tahun'=>$tahun])
 							->count_all_results();
 		}
 
 		public function getById($id, $pengiriman=false) {
-			$this->db->from('pengiriman_akuntansi')
-					->join('permintaan_akuntansi', 'permintaan_akuntansi.id_permintaan = pengiriman_akuntansi.id_permintaan', 'left')
-					->join('jenis_data', 'permintaan_akuntansi.kode_jenis = jenis_data.kode_jenis', 'left')
-					->join('klien', 'permintaan_akuntansi.id_klien = klien.id_klien', 'left')
-					->join('tugas', 'jenis_data.kode_jenis = tugas.kode_jenis AND klien.status_pekerjaan = tugas.status_pekerjaan', 'left');
-			if($pengiriman == true) {
-				$this->db->where(['id_pengiriman'=>$id]);
-			} else {
-				$this->db->join('proses_akuntansi', 'proses_akuntansi.id_kirim = pengiriman_akuntansi.id_pengiriman AND tugas.id_tugas = proses_akuntansi.id_tugas', 'left')
-						->where(['id_proses'=>$id]);
-			}
-			return $this->db->get()->row_array();
+			return $this->db->from('data_akuntansi')
+							->join('proses_akuntansi', 'proses_akuntansi.id_proses = data_akuntansi.id_kerja', 'left')
+							->join('permintaan_akuntansi', 'permintaan_akuntansi.id_permintaan = data_akuntansi.id_request', 'left')
+							->join('pengiriman_akuntansi', 'pengiriman_akuntansi.id_pengiriman = data_akuntansi.id_kirim', 'left')
+							->join('jenis_data', 'jenis_data.kode_jenis = data_akuntansi.id_jenis', 'left')
+							->join('klien', 'klien.id_klien = permintaan_akuntansi.id_klien', 'left')
+							->join('tugas', 'tugas.id_jenis = jenis_data.kode_jenis AND tugas.status_pekerjaan = klien.status_pekerjaan', 'left')
+							->join('user', 'user.id_user = proses_akuntansi.id_akuntan', 'left')
+							->where(['id_proses' => $id])
+							->get()->row_array();
 		}
 		
 		public function tambahProses() {
-			$id_pengiriman	= $this->input->post('id_pengiriman', true);
-			$id_akuntan		= $this->input->post('id_akuntan', true);
-			$id_tugas		= $this->input->post('id_tugas', true);
-			$mulai			= $this->input->post('tanggal_mulai', true).' '.$this->input->post('jam_mulai', true);
-			$selesai		= $this->input->post('tanggal_selesai', true).' '.$this->input->post('jam_selesai', true);
+			$mulai		= $this->input->post('tanggal_mulai', true).' '.$this->input->post('jam_mulai', true);
+			$selesai	= $this->input->post('tanggal_selesai', true).' '.$this->input->post('jam_selesai', true);
 			
-			$id_tugas	= substr($id_tugas, 3);
-			$id_proses	= $id_pengiriman . $id_tugas . $id_akuntan;
-			
-			$flag = 0;
-			if( $this->input->post('tanggal_selesai', true) ) {
-				if($this->input->post('jam_selesai', true) == null) {
-					$this->session->set_flashdata('jam_selesai', '<b>Jam Selesai</b> harus diisi');
-					redirect('akuntan/proses_data_akuntansi/mulai/'.$id_pengiriman);
-				} else {
-					$flag = 1;
-				}
-			} else {
-				$selesai = null;
-				if($this->input->post('jam_selesai', true) != null) {
-					$this->session->set_flashdata('tanggal_selesai', '<b>Tanggal Selesai</b> harus diisi');
-					redirect('akuntan/proses_data_akuntansi/mulai/'.$id_pengiriman);
-				} else {
-					$flag = 1;
-				}
-			}
-
-			if($flag == 1) {
-				$data = [
-					'id_proses'			=> $id_proses,
-					'tanggal_proses'	=> date('d-m-Y H:i'),
-					'tanggal_mulai'		=> $mulai,
-					'tanggal_selesai'	=> $selesai,
-					'keterangan3'		=> $this->input->post('keterangan3', true),
-					'id_tugas'			=> $this->input->post('id_tugas', true),
-					'id_kirim'			=> $this->input->post('id_pengiriman', true),
-					'id_akuntan'		=> $this->input->post('id_akuntan', true),
-				];
-				$this->db->insert('proses_akuntansi', $data);
-			}
+			$data = [
+				'tanggal_proses'	=> date('d-m-Y H:i'),
+				'tanggal_mulai'		=> $mulai,
+				'tanggal_selesai'	=> ($selesai == ' ') ? null : $selesai,
+				'ket_proses'		=> $this->input->post('keterangan3', true),
+				'id_akuntan'		=> $this->session->userdata('id_user'),
+			];
+			$this->db->where('id_proses', $this->input->post('id_proses', true))
+					->update('proses_akuntansi', $data);
 		}
 		
 		public function ubahProses() {
 			$data = [
-				//'tanggal_proses'	=> date('d-m-Y H:i'),
-				'tanggal_mulai'		=> $this->input->post('tanggal_mulai', true).' '.$this->input->post('jam_mulai', true),
 				'tanggal_selesai'	=> $this->input->post('tanggal_selesai', true).' '.$this->input->post('jam_selesai', true),
-				'keterangan3'		=> $this->input->post('keterangan3', true),
+				'ket_proses'		=> $this->input->post('keterangan3', true),
 			];
-			$this->db->where('id_proses', $this->input->post('id_proses', true));
-			$this->db->update('proses_akuntansi', $data);
+			$this->db->where('id_proses', $this->input->post('id_proses', true))
+					->update('proses_akuntansi', $data);
 		}
-		
-		public function batalMulaiProses($data) {
+
+		public function batalProses($data) {
 			$max = $this->db->select_max('idt_proses')
 							->where('id_proses', $data['id_proses'])
 							->get('trash_proses_akuntansi')->row_array();
@@ -143,55 +98,29 @@
 			} else {
 				$idt_proses	= $data['id_proses'] . '00';
 			}
-
-			$row = [
-				'idt_proses'		=> $idt_proses,
-				'tanggal_cancel'	=> date('d-m-Y H:i'),
-				'id_proses'			=> $data['id_proses'],
-				'tanggal_proses'	=> $data['tanggal_proses'],
-				'tanggal_mulai'		=> $data['tanggal_mulai'],
-				'tanggal_selesai'	=> $data['tanggal_selesai'],
-				'keterangan3'		=> $data['keterangan3'],
-				'id_tugas'			=> $data['id_tugas'],
-				'id_kirim'			=> $data['id_kirim'],
-				'id_akuntan'		=> $data['id_akuntan'],
-				'id_disposer3'		=> $data['id_disposer3'],
-			];
-			$this->db->insert('trash_proses_akuntansi', $row);
 			
-			$this->db->delete('proses_akuntansi', ['id_proses' => $data['id_proses']]);
-			$this->session->set_flashdata('notification', 'Proses berhasil dibatalkan!'); 
-		}
-
-		public function batalSelesaiProses($data) {
-			$max = $this->db->select_max('idt_proses')
-							->where('id_proses', $data['id_proses'])
-							->get('trash_proses_akuntansi')->row_array();
-			if($max) {
-				$idt		= substr($max['idt_proses'], -2);
-				$idt_proses	= $data['id_proses'] . ++$idt;
-			} else {
-				$idt_proses	= $data['id_proses'] . '00';
-			}
-
-			$row = [
+			$trow = [
 				'idt_proses'		=> $idt_proses,
 				'tanggal_cancel'	=> date('d-m-Y H:i'),
 				'id_proses'			=> $data['id_proses'],
 				'tanggal_proses'	=> $data['tanggal_proses'],
 				'tanggal_mulai'		=> $data['tanggal_mulai'],
 				'tanggal_selesai'	=> $data['tanggal_selesai'],
-				'keterangan3'		=> $data['keterangan3'],
-				'id_tugas'			=> $data['id_tugas'],
-				'id_kirim'			=> $data['id_kirim'],
-				'id_akuntan'		=> $data['id_akuntan'],
+				'ket_proses'		=> $data['ket_proses'],
+				'id_data'			=> $data['id_data'],
 				'id_disposer3'		=> $data['id_disposer3'],
 			];
-			$this->db->insert('trash_proses_akuntansi', $row);
-
-			$this->db->where(['id_proses' => $data['id_proses']])
-					->update('proses_akuntansi', ['tanggal_selesai' => null]);
-			$this->session->set_flashdata('notification', 'Proses berhasil dibatalkan!'); 
+			$this->db->insert('trash_proses_akuntansi', $trow);
+			
+			if($data['jenis_proses'] == 'onproses') {
+				$this->db->delete('proses_akuntansi', ['id_proses' => $data['id_proses']]);
+			} else {
+				$row = [
+					'tanggal_selesai'	=> null,
+					'temp_selesai'		=> $data['tanggal_selesai'],
+				];
+				$this->db->update('proses_akuntansi', $row, ['id_proses' => $data['id_proses']]);
+			}
 		}
 	}
 ?>
